+++ cider-mode.el
@@ -41,6 +41,7 @@
     (define-key map (kbd "C-c C-c") 'cider-eval-expression-at-point)
     (define-key map (kbd "C-x C-e") 'cider-eval-last-expression)
     (define-key map (kbd "C-c C-e") 'cider-eval-last-expression)
+    (define-key map (kbd "C-c M-e") 'cider-eval-last-expression-to-repl)
     (define-key map (kbd "C-c C-p") 'cider-pprint-eval-last-expression)
     (define-key map (kbd "C-c C-r") 'cider-eval-region)
     (define-key map (kbd "C-c C-n") 'cider-eval-ns-form)